Finally fixed the local variable pushing problem
Fixing the `return` keyword allowed to fix the local variable problem.
Before, `let` variables were being converted to real because if we did not do that, the compiler would create a very weird segfault.
We are now keeping the `let` variables as `@localconst` items which allows us to remove more than 20k lines from the assembly. Nice!!!